WLT-998: change initialization.go generate
