+++ insolar/rootdomain/rootdomain_test.go
@@ -17,35 +17,12 @@
 package rootdomain
 
 import (
-	"encoding/hex"
 	"testing"
 
 	"github.com/insolar/insolar/insolar/genesisrefs"
 
-	"github.com/insolar/insolar/insolar"
-	"github.com/insolar/insolar/platformpolicy"
 	"github.com/stretchr/testify/require"
 )
-
-var (
-	idHex  = "00010001c896b5c98f56001c688bc80a48274ac266a780a5a7ae74c4e1e3624b"
-	refHex = idHex + idHex
-)
-
-func TestID(t *testing.T) {
-	rootRecord := &Record{
-		PCS: initPCS(),
-	}
-	require.Equal(t, idHex, hex.EncodeToString(rootRecord.ID().Bytes()), "root domain ID should always be the same")
-}
-
-func TestReference(t *testing.T) {
-	rootRecord := &Record{
-		PCS: initPCS(),
-	}
-	require.Equal(t, refHex, hex.EncodeToString(rootRecord.Reference().Bytes()), "root domain Ref should always be the same")
-
-}
 
 func TestGenesisRef(t *testing.T) {
 	var (
@@ -54,8 +31,4 @@ func TestGenesisRef(t *testing.T) {
 	)
 	genesisRef := genesisrefs.GenesisRef(pubKey)
 	require.Equal(t, pubKeyRef, genesisRef.String(), "reference by name always the same")
-}
-
-func initPCS() insolar.PlatformCryptographyScheme {
-	return platformpolicy.NewPlatformCryptographyScheme()
 }
